+++ product-server/src/main/java/ru/itmp/productserver/controller/CategoryController.java
@@ -5,14 +5,15 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
-//import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 import ru.itmp.productserver.dto.request.CategoryRequest;
 import ru.itmp.productserver.dto.responce.CategoryResponse;
 import ru.itmp.productserver.dto.update.CategoryUpdate;
 import ru.itmp.productserver.services.CategoryService;
-
+import ru.itmp.productserver.feign.AuthFeignClient;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
 
 import javax.validation.Valid;
 import java.util.List;
@@ -23,41 +24,48 @@ import java.util.List;
 @RequestMapping(value = "/api/v1/category")
 public class CategoryController {
     private final CategoryService categoryService;
+    private final AuthFeignClient authFeign;
 
     @PostMapping
-    @PreAuthorize("hasAnyAuthority('ADMIN')")
-    public CategoryResponse addCategory(@Valid @RequestBody CategoryRequest categoryRequest) {
-        return categoryService.save(categoryRequest);
+    public ResponseEntity<CategoryResponse> addCategory(@RequestBody CategoryRequest categoryRequest, @RequestHeader("Authorization") String token) {
+        if (!authFeign.checkAdminPermission(token)) {
+            System.out.println("checkAdminPermission: false");
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(null);
+        }
+        return new ResponseEntity<>(categoryService.save(categoryRequest), HttpStatus.CREATED);
     }
 
     @GetMapping
-    @PreAuthorize("hasAnyAuthority('ADMIN', 'USER')")
-    public Page<CategoryResponse> getListCategory(@PageableDefault(size = 5) Pageable pageable) {
-        return categoryService.getAllPage(pageable);
+    public ResponseEntity<Page<CategoryResponse>> getListCategory(@PageableDefault(size = 5) Pageable pageable) {
+        return ResponseEntity.ok(categoryService.getAllPage(pageable));
     }
 
     @GetMapping("/all")
-    @PreAuthorize("hasAnyAuthority('ADMIN', 'USER')")
-    public List<CategoryResponse> getAllCategory() {
-        return categoryService.getAllList();
+    public ResponseEntity<List<CategoryResponse>> getAllCategory() {
+        return ResponseEntity.ok(categoryService.getAllList());
     }
 
     @PutMapping
-    @PreAuthorize("hasAnyAuthority('ADMIN')")
-    public CategoryResponse updateCategory(@Valid @RequestBody CategoryUpdate categoryUpdate) {
-        return categoryService.update(categoryUpdate);
+    public ResponseEntity<CategoryResponse> updateCategory(@RequestBody CategoryUpdate categoryUpdate, @RequestHeader("Authorization") String token) {
+        if (!authFeign.checkAdminPermission(token)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(null);
+        }
+        return new ResponseEntity<>(categoryService.update(categoryUpdate), HttpStatus.OK);
     }
 
     @GetMapping("/{id}")
-    @PreAuthorize("hasAnyAuthority('ADMIN', 'USER')")
-    public CategoryResponse getCategoryDetail(@PathVariable Long id) {
-        return categoryService.getCategoryById(id);
+    public ResponseEntity<CategoryResponse> getCategoryDetail(@PathVariable Long id) {
+        return ResponseEntity.ok(categoryService.getCategoryById(id));
     }
 
     @DeleteMapping("/{id}")
-    @PreAuthorize("hasAnyAuthority('ADMIN')")
-    public void deleteCategory(@PathVariable Long id) {
+    public ResponseEntity<String> deleteCategory(@PathVariable Long id, @RequestHeader("Authorization") String token) {
+        if (!authFeign.checkAdminPermission(token)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(null);
+        }
         categoryService.deleteById(id);
+        return ResponseEntity.ok("Category deleted successfully");
+
     }
 
 }
